Добавил документацию и попытался сделать массив динамическим. Пока неуспешно.
